Create
The Gaming Room wanted me to develop a web-based game that serves multiple platforms based on their current game, Draw It or Lose It, which was currently available in an Android app only. Draw It or Lose It is loosely similar to the 1980s television game Win, Lose or Draw, where teams compete to guess what is being drawn. Rather than a player drawing images on an easel to help team members guess the puzzle (a phrase, title, or thing), the application will render images from a large library of stock drawings as clues. A game consists of four rounds of play lasting one minute each. Drawings are rendered at a steady rate and are fully complete at the 30-second mark. If the team does not guess the puzzle before time expires, the remaining teams have an opportunity to offer one guess each to solve the puzzle with a 15-second time limit. 

What I did well in creating this documentation is making sure to follow the rubric and include all the needed parts and more. I made sure to include details about the client and server-side as well as development tools. What I could have done better is my executive summary, since it was a bit different then what was needed. Working on the design document really helped me get a better understanding of the software and how I would design it and what I would need. I would revise my executive summary if I were to choose on thing to revise, and make it a bit more concentrated on the topic. I knew that the user needed to be able to work on multiple platforms and operating systems so I chose a client-server architecture which supports this through the use of a single backend REST API that any client can communicate with. In the future I would like to create documents like these when developing complex software so that I have a vision of what is needed before I start coding. Additionaly flowcharts and UML diagrams help too.